+++ App.xaml.cs
@@ -1,6 +1,4 @@
-﻿using System.Configuration;
-using System.Data;
-using System.Windows;
+﻿using System.Windows;
 
 namespace Tfx;
 
@@ -9,5 +7,19 @@ namespace Tfx;
 /// </summary>
 public partial class App : Application
 {
+    protected override void OnStartup(StartupEventArgs e)
+    {
+        base.OnStartup(e);
+
+        var options = StartupOptions.Parse(e.Args);
+        if (options.ShowHelp || options.HasError)
+        {
+            options.WriteHelp();
+            Shutdown(options.HasError ? 1 : 0);
+            return;
+        }
+
+        new MainWindow(options).Show();
+    }
 }